feat: refactor conditional compilation using split module pattern (#171)
Major refactoring to reduce conditional compilation complexity by 35%:

- Applied split module pattern to isolate async/blocking code
- Reduced cfg attributes from ~250 to 162 (35% reduction)
- Each file now compiles entirely under one feature configuration
- Added comprehensive developer documentation

Module-specific improvements:
- socket_manager: 78% cfg reduction, split into tokio.rs/mod.rs
- channels: 65% cfg reduction, split into tokio.rs/mod.rs
- executor: 90% cfg reduction, split into async_impl/blocking/mod.rs
- camera/methods: 88% cfg reduction in reexports
- camera/generic: 41% cfg reduction, split into mod.rs/async_impl/blocking

Additional changes:
- Added 'static bounds to async trait implementations (from #172)
- Added #\![deny(unreachable_code, dead_code)] for code quality
- Created follow-up issue #173 for automated lint enforcement
- All tests pass in both blocking and async modes

This refactoring significantly improves code maintainability while
preserving full API compatibility (except intended 'static bounds).
