+++ src/camera/generic/mod.rs
@@ -1,0 +1,421 @@
+//! Generic camera implementation using compile-time profiles.
+//!
+//! This module provides the new generic Camera<P, T> struct that uses
+//! compile-time profile selection for zero-cost abstractions.
+
+use std::marker::PhantomData;
+use std::sync::Arc;
+use std::time::Duration;
+
+use crate::{
+    camera_id::CameraId,
+    capabilities::{CameraFeature, FeatureDetection, Profile},
+    command::{encode_visca::EncodeVisca, Response},
+    error::Error,
+    transport::{TransportEnvelope, UnifiedTransport},
+};
+
+#[cfg(feature = "async")]
+use crate::socket_manager::SocketManagerHandle;
+
+#[cfg(feature = "async")]
+use crate::executor::Spawner;
+
+#[cfg(feature = "async")]
+mod async_impl;
+
+mod blocking;
+
+/// Generic camera client with compile-time profile selection.
+///
+/// This struct provides type-safe camera control with zero runtime overhead.
+/// All profile-specific constants and behaviors are resolved at compile time.
+///
+/// # Type Parameters
+///
+/// * `P` - Camera profile implementing the `Profile` trait
+/// * `T` - Transport implementing `UnifiedTransport`
+///
+/// # Examples
+///
+/// ```ignore
+/// use grafton_visca::camera::generic::Camera;
+/// use grafton_visca::camera::profiles::PTZOpticsG2;
+/// use grafton_visca::transport::blocking::Tcp;
+///
+/// // Create a camera with explicit profile type
+/// let transport = Tcp::connect("192.168.1.100:52381")?;
+/// let camera = Camera::<PTZOpticsG2, _>::new(transport);
+///
+/// // Or use the type alias
+/// use grafton_visca::prelude::PTZOpticsG2Cam;
+/// let camera = PTZOpticsG2Cam::new(transport);
+/// ```
+pub struct Camera<P, T>
+where
+    P: Profile,
+    T: UnifiedTransport,
+{
+    transport: Arc<T>,
+    camera_id: CameraId,
+    #[cfg(feature = "async")]
+    socket_manager: Option<SocketManagerHandle>,
+    envelope: TransportEnvelope,
+    #[cfg(feature = "async")]
+    spawner: Option<Arc<dyn Spawner>>,
+    _profile: PhantomData<P>,
+}
+
+impl<P, T> Clone for Camera<P, T>
+where
+    P: Profile,
+    T: UnifiedTransport,
+{
+    fn clone(&self) -> Self {
+        Self {
+            transport: Arc::clone(&self.transport),
+            camera_id: self.camera_id,
+            #[cfg(feature = "async")]
+            socket_manager: self.socket_manager.clone(),
+            envelope: TransportEnvelope::new(P::PROTOCOL_STYLE),
+            #[cfg(feature = "async")]
+            spawner: self.spawner.clone(),
+            _profile: PhantomData,
+        }
+    }
+}
+
+impl<P, T> std::fmt::Debug for Camera<P, T>
+where
+    P: Profile,
+    T: UnifiedTransport,
+{
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        let mut debug = f.debug_struct("Camera");
+        debug
+            .field("profile", &P::MODEL_NAME)
+            .field("camera_id", &self.camera_id)
+            .field("transport", &"<UnifiedTransport>");
+        #[cfg(feature = "async")]
+        debug.field("socket_manager", &self.socket_manager.is_some());
+        debug.finish()
+    }
+}
+
+impl<P, T> Camera<P, T>
+where
+    P: Profile,
+    T: UnifiedTransport,
+{
+    /// Create a new camera from a unified transport.
+    pub fn from_transport(transport: T) -> Self {
+        let camera_id = CameraId::new(P::DEFAULT_ADDRESS).unwrap_or(CameraId::CAMERA_1);
+
+        Self {
+            envelope: TransportEnvelope::new(P::PROTOCOL_STYLE),
+            transport: Arc::new(transport),
+            camera_id,
+            #[cfg(feature = "async")]
+            socket_manager: None,
+            #[cfg(feature = "async")]
+            spawner: None,
+            _profile: PhantomData,
+        }
+    }
+
+    /// Get the camera's model name.
+    #[must_use]
+    pub fn model_name(&self) -> &'static str {
+        P::MODEL_NAME
+    }
+
+    /// Set the camera ID for this camera.
+    ///
+    /// # Errors
+    /// Returns an error if the ID is not in the valid range (1-8).
+    pub fn set_camera_id(&mut self, id: u8) -> Result<(), Error> {
+        self.camera_id = CameraId::new(id)?;
+        Ok(())
+    }
+
+    /// Get the current camera ID.
+    #[must_use]
+    pub fn camera_id(&self) -> CameraId {
+        self.camera_id
+    }
+
+    /// Get profile-specific constants directly.
+    ///
+    /// # Examples
+    ///
+    /// ```ignore
+    /// let zoom_range = camera.zoom_speed_range();
+    /// let max_pan_speed = camera.max_pan_speed();
+    /// ```
+    #[must_use]
+    pub fn zoom_speed_range(&self) -> std::ops::Range<u8> {
+        P::ZOOM_SPEED_RANGE
+    }
+
+    /// Get the maximum pan speed.
+    #[must_use]
+    pub fn max_pan_speed(&self) -> u8 {
+        P::MAX_PAN_SPEED
+    }
+
+    /// Get the maximum tilt speed.
+    #[must_use]
+    pub fn max_tilt_speed(&self) -> u8 {
+        P::MAX_TILT_SPEED
+    }
+
+    /// Get the maximum optical zoom value.
+    #[must_use]
+    pub fn optical_zoom_max(&self) -> u16 {
+        P::OPTICAL_ZOOM_MAX
+    }
+
+    /// Get the maximum digital zoom value.
+    #[must_use]
+    pub fn digital_zoom_max(&self) -> Option<u16> {
+        P::DIGITAL_ZOOM_MAX
+    }
+
+    /// Get the maximum number of presets.
+    #[must_use]
+    pub fn max_presets(&self) -> u8 {
+        P::MAX_PRESETS
+    }
+
+    /// Get the power on time duration.
+    #[must_use]
+    pub fn power_on_time(&self) -> Duration {
+        P::POWER_ON_TIME
+    }
+
+    /// Get the standby time duration.
+    #[must_use]
+    pub fn standby_time(&self) -> Duration {
+        P::STANDBY_TIME
+    }
+
+    /// Convert degrees to pan/tilt units.
+    #[must_use]
+    pub fn degrees_to_units(
+        &self,
+        pan_deg: crate::units::Degrees,
+        tilt_deg: crate::units::Degrees,
+    ) -> (crate::units::ViscaUnits<i16>, crate::units::ViscaUnits<i16>) {
+        let pan_units = (pan_deg.0 * P::PAN_DEGREES_TO_UNITS) as i16;
+        let tilt_units = (tilt_deg.0 * P::TILT_DEGREES_TO_UNITS) as i16;
+
+        (
+            crate::units::ViscaUnits(pan_units),
+            crate::units::ViscaUnits(tilt_units),
+        )
+    }
+
+    /// Convert pan/tilt units to degrees.
+    #[must_use]
+    pub fn units_to_degrees(
+        &self,
+        pan_units: i16,
+        tilt_units: i16,
+    ) -> (crate::units::Degrees, crate::units::Degrees) {
+        let pan_deg = pan_units as f32 / P::PAN_DEGREES_TO_UNITS;
+        let tilt_deg = tilt_units as f32 / P::TILT_DEGREES_TO_UNITS;
+
+        (
+            crate::units::Degrees(pan_deg),
+            crate::units::Degrees(tilt_deg),
+        )
+    }
+
+    /// Validate pan position.
+    pub fn validate_pan(&self, pan_units: i16) -> Result<i16, Error> {
+        use crate::capabilities::ValidationError;
+
+        if P::PAN_RANGE.contains(&pan_units) {
+            Ok(pan_units)
+        } else {
+            Err(Error::ValidationError(ValidationError::OutOfRange {
+                parameter: "pan",
+                value: pan_units as f64,
+                min: P::PAN_RANGE.start as f64,
+                max: P::PAN_RANGE.end as f64,
+            }))
+        }
+    }
+
+    /// Validate tilt position.
+    pub fn validate_tilt(&self, tilt_units: i16) -> Result<i16, Error> {
+        use crate::capabilities::ValidationError;
+
+        if P::TILT_RANGE.contains(&tilt_units) {
+            Ok(tilt_units)
+        } else {
+            Err(Error::ValidationError(ValidationError::OutOfRange {
+                parameter: "tilt",
+                value: tilt_units as f64,
+                min: P::TILT_RANGE.start as f64,
+                max: P::TILT_RANGE.end as f64,
+            }))
+        }
+    }
+
+    /// Check if a feature is supported by this camera profile.
+    #[must_use]
+    pub fn has_feature(&self, feature: CameraFeature) -> bool {
+        // For now, we can't determine feature support at compile time
+        // without a ProfileIntrospection trait that profiles can implement
+        self.supports_feature(feature)
+    }
+
+    /// Get a reference to the transport.
+    #[must_use]
+    pub fn transport(&self) -> &Arc<T> {
+        &self.transport
+    }
+
+    /// Get the camera's coordinate system.
+    #[must_use]
+    pub fn coordinate_system(&self) -> crate::capabilities::CoordinateSystem {
+        P::COORDINATE_SYSTEM
+    }
+
+    /// Convert logical pan/tilt coordinates to camera-specific coordinates.
+    #[must_use]
+    pub fn to_camera_coords(&self, pan: i16, tilt: i16) -> (u16, u16) {
+        self.coordinate_system().to_camera_coords(pan, tilt)
+    }
+
+    /// Convert camera-specific coordinates to logical pan/tilt coordinates.
+    #[must_use]
+    pub fn from_camera_coords(&self, pan: u16, tilt: u16) -> (i16, i16) {
+        self.coordinate_system()
+            .convert_from_camera_coords(pan, tilt)
+    }
+}
+
+/// Extension trait for feature detection.
+impl<P, T> FeatureDetection for Camera<P, T>
+where
+    P: Profile,
+    T: UnifiedTransport,
+{
+    fn supports_feature(&self, feature: CameraFeature) -> bool {
+        // For compile-time profiles, we check against known mandatory features
+        match feature {
+            CameraFeature::PanTilt => true, // All profiles must implement PanTilt
+            CameraFeature::Zoom => true,    // All profiles must implement Zoom
+            CameraFeature::Focus => true,   // All profiles must implement Focus
+            CameraFeature::Exposure => true, // All profiles must implement Exposure
+            CameraFeature::WhiteBalance => true, // All profiles must implement WhiteBalance
+            CameraFeature::ImageProcessing => true, // All profiles must implement ImageProcessing
+            CameraFeature::Power => true,   // All profiles must implement Power
+            CameraFeature::Presets => true, // All profiles must implement Presets
+            CameraFeature::MenuControl => true, // All profiles must implement MenuControl
+            // Optional features - we can't detect these at runtime with the current design
+            _ => false,
+        }
+    }
+
+    fn supports_command(&self, _command: &dyn std::any::Any) -> bool {
+        // For now, we can't determine command support without CommandFeatures trait
+        // This would require all commands to implement CommandFeatures
+        // Return true for backward compatibility
+        true
+    }
+
+    fn supported_features(&self) -> Vec<CameraFeature> {
+        // Return the features that all profiles must support
+        vec![
+            CameraFeature::PanTilt,
+            CameraFeature::Zoom,
+            CameraFeature::Focus,
+            CameraFeature::Exposure,
+            CameraFeature::WhiteBalance,
+            CameraFeature::ImageProcessing,
+            CameraFeature::Power,
+            CameraFeature::Presets,
+            CameraFeature::MenuControl,
+        ]
+    }
+}
+
+/// Command execution implementation.
+#[cfg(not(feature = "async"))]
+impl<P, T> Camera<P, T>
+where
+    P: Profile,
+    T: UnifiedTransport,
+{
+    /// Send a VISCA command to the camera without expecting a response.
+    pub fn send_command_no_response<C: EncodeVisca>(&self, command: &C) -> Result<(), Error> {
+        let mut buffer = [0u8; 64];
+        let size = command.encode_into(self.camera_id, &mut buffer)?;
+        let cmd_bytes = &buffer[..size];
+
+        // Apply protocol-specific framing
+        let framed_bytes = self.envelope.frame_command(cmd_bytes, false);
+        self.transport.send_blocking(&framed_bytes)?;
+        Ok(())
+    }
+
+    /// Send a VISCA command to the camera and wait for a response.
+    pub fn send_command_sync<C: EncodeVisca>(
+        &self,
+        command: &C,
+        _expected_response: crate::command::ResponseType,
+    ) -> Result<Response, Error> {
+        // Use the blocking method
+        self.send_command_blocking(command)
+    }
+
+    /// Send raw bytes to the camera.
+    ///
+    /// # Safety
+    /// This bypasses all validation and can send invalid commands.
+    /// Use with caution and only when implementing custom commands.
+    pub fn send_raw(&self, data: &[u8]) -> Result<(), Error> {
+        self.transport.send_blocking(data)
+    }
+}
+
+/// Command execution implementation.
+impl<P, T> Camera<P, T>
+where
+    P: Profile,
+    T: UnifiedTransport,
+{
+    /// Send a command to the camera using blocking method.
+    /// This is used internally by the blocking trait implementations.
+    pub(crate) fn send_command_blocking<C>(&self, command: &C) -> Result<Response, Error>
+    where
+        C: EncodeVisca,
+    {
+        // Always use the blocking implementation
+        self.send_command_blocking_impl(command)
+    }
+}
+
+// Impl block for convenience constructor with Transport types
+#[cfg(not(feature = "async"))]
+impl<P, AT> Camera<P, crate::transport::AsyncTransportWrapper<AT>>
+where
+    P: Profile,
+    AT: crate::transport::core::Transport + Send + Sync + 'static,
+    for<'a> AT::SendFut<'a>: Send,
+    for<'a> AT::RecvFut<'a>: Send,
+{
+    /// Create a new camera with a transport implementing the Transport trait.
+    ///
+    /// This is a convenience method for blocking mode.
+    pub fn new(transport: AT) -> Self {
+        let wrapped = crate::transport::AsyncTransportWrapper { transport };
+        Self::from_transport(wrapped)
+    }
+}
+
+// Re-export Camera type at module level for backward compatibility
+pub use Camera as GenericCamera;
